+++ tests/integration/standard/test_query.py
@@ -191,7 +191,7 @@ class PrintStatementTests(unittest.TestCase):
         Highlight the format of printing SimpleStatements
         """
 
-        ss = SimpleStatement('SELECT * FROM test3rf.test')
+        ss = SimpleStatement('SELECT * FROM test3rf.test', consistency_level=ConsistencyLevel.ONE)
         self.assertEqual(str(ss),
                          '<SimpleStatement query="SELECT * FROM test3rf.test", consistency=ONE>')
 
@@ -204,6 +204,7 @@ class PrintStatementTests(unittest.TestCase):
         session = cluster.connect()
 
         prepared = session.prepare('INSERT INTO test3rf.test (k, v) VALUES (?, ?)')
+        prepared.consistency_level = ConsistencyLevel.ONE
 
         self.assertEqual(str(prepared),
                          '<PreparedStatement query="INSERT INTO test3rf.test (k, v) VALUES (?, ?)", consistency=ONE>')